+++ examples/boxes/main.cpp
@@ -39,6 +39,7 @@
 #include <osgShadow/ShadowMap>
 
 using namespace dart;
+using dart::simulation::CollisionDetectorType;
 
 [[nodiscard]] dynamics::SkeletonPtr createBox(
     const Eigen::Vector3d& position,
@@ -78,8 +79,7 @@ int main()
   auto world = simulation::World::create();
 
   // Set collision detector type
-  world->getConstraintSolver()->setCollisionDetector(
-      collision::BulletCollisionDetector::create());
+  world->setCollisionDetector(collision::BulletCollisionDetector::create());
 
   // Create dim x dim x dim boxes
   auto dim = 5;