+++ examples/fetch/main.cpp
@@ -41,6 +41,7 @@
 #include <dart/All.hpp>
 
 using namespace dart;
+using dart::simulation::CollisionDetectorType;
 
 class FetchWorldNode : public gui::osg::RealTimeWorldNode
 {
@@ -157,8 +158,7 @@ int main()
   auto world = utils::MjcfParser::readWorld(
       "dart://sample/mjcf/openai/robotics/fetch/pick_and_place.xml");
   DART_ASSERT(world);
-  world->getConstraintSolver()->setCollisionDetector(
-      collision::BulletCollisionDetector::create());
+  world->setCollisionDetector(CollisionDetectorType::Bullet);
 
   // Get Fetch robot and set properties
   auto robot = world->getSkeleton("robot0:base_link");